0.0.0.10
아이템 추가,
아이템 습득시 인벤토리에 표기

+++ unity_project/Assets/Scripts/Item.cs
@@ -20,6 +20,6 @@ public class Item : ScriptableObject
     public Sprite       itemImage;          // 아이템 이미지
     public GameObject   itemPrefab;         // 아이템 객체
 
-    public float        itemAmount;         // 아이템 갯수
-    public float        itemMaxAmount;      // 아이템 최대 갯수
+    public int          itemAmount;         // 아이템 갯수
+    public int          itemMaxAmount;      // 아이템 최대 갯수
 }
